SystemUI: fix qstile page reset timeout
When the user hasn't visisted the quick settings tile area for some time
(currently 10s), we need to reset the page back to the first one.

We can't go off whether we are listening to events because when the user
turns the screen off, listening is set to true, so the page would never
reset if the user turned their screen off.

We now key off whether we are expanded, which represents whether the
user is interacting with the panel.

Change-Id: I010aaa9e9ea84eb96308184dad606d47ce136748
Signed-off-by: Roman Birg <roman@cyngn.com>